Move all robot files to separate package
